+++ aldryn_forms/sizefield/utils.py
@@ -1,11 +1,9 @@
-from __future__ import unicode_literals
-
 import operator
 import re
 import sys
 
 from django.conf import settings
-from django.utils import formats, six
+from django.utils import formats
 from django.utils.translation import ugettext as _
 
 
@@ -43,7 +41,7 @@ def filesizeformat(bytes, decimals=1):
     def filesize_number_format(value):
         return formats.number_format(round(value, decimals), decimals)
 
-    units_list = sorted(six.iteritems(FILESIZE_UNITS), key=operator.itemgetter(1))
+    units_list = sorted(FILESIZE_UNITS.items(), key=operator.itemgetter(1))
 
     value = unit = None
     len_unints_list = len(units_list)
@@ -65,7 +63,7 @@ def parse_size(size):
     """
     @rtype int
     """
-    if isinstance(size, six.integer_types):
+    if isinstance(size, int):
         return size
 
     r = file_size_re.match(size.strip())